Revert "Move audio resampler CPU implementation to a single compilation unit (#3914)"
This reverts commit 6c844cc4bcb23df21bc37e73a9dcce602b2cf4a8.
