+++ dali/operators/decoder/audio/audio_decoder_impl.h
@@ -19,7 +19,7 @@
 #include "dali/operators/decoder/audio/audio_decoder.h"
 #include "dali/operators/decoder/audio/generic_decoder.h"
 #include "dali/pipeline/data/backend.h"
-#include "dali/kernels/signal/resampling_cpu.h"
+#include "dali/kernels/signal/resampling.h"
 #include "dali/core/tensor_view.h"
 
 namespace dali {
@@ -42,7 +42,7 @@ DLL_PUBLIC std::pair<int64_t, int64_t> ProcessOffsetAndLength(const AudioMetadat
  * @param target_sample_rate If a positive number is provided, it represent the target sampling rate
  *                           (the audio data is expected to be resampled if its original sampling rate differs)
  * @param downmix If set to true, the audio channels are expected to be downmixed, resulting in a shape with 1
- *                dimension ({nsamples,}), instead of 2 ({nsamples, nchannels})
+ *                dimension ({nsamples,}), instead of 2 ({nsamples, nchannels})       
  */
 DLL_PUBLIC TensorShape<> DecodedAudioShape(const AudioMetadata &meta, float target_sample_rate = -1,
                                            bool downmix = true);
@@ -52,23 +52,23 @@ DLL_PUBLIC TensorShape<> DecodedAudioShape(const AudioMetadata &meta, float targ
  * @param audio Destination buffer. The function will decode as many audio samples as the shape of this argument
  * @param decoder Decoder object.
  * @param meta Audio metadata.
- * @param resampler ResamplerCPU instance used if resampling is required
+ * @param resampler Resampler instance used if resampling is required
  * @param decode_scratch_mem Scratch memory used for decoding, when decoding can't be done directly to the output buffer.
  *                           If downmixing or resampling is required, this buffer should have a positive length, representing
  *                           decoded audio length at the original sampling rate: ``length * nchannels``
  * @param resample_scratch_mem Scratch memory used for the input of resampling.
- *                             If resampling is required, the buffer should have a positive length, representing the
+ *                             If resampling is required, the buffer should have a positive length, representing the 
  *                             decoded audio length, ``length`` if downmixing is enabled, or the decoded audio length including
  *                             channels, ``length * nchannels``, otherwise.
  * @param target_sample_rate If a positive value is provided, the signal will be resampled except when its original sampling rate
  *                           is equal to the target.
  * @param downmix If true, the audio channes will be downmixed to a single one
- * @param audio_filepath Path to the audio file being decoded, only used for debugging purposes
+ * @param audio_filepath Path to the audio file being decoded, only used for debugging purposes 
  */
 template <typename T>
 DLL_PUBLIC void DecodeAudio(TensorView<StorageCPU, T, DynamicDimensions> audio,
                             AudioDecoderBase &decoder, const AudioMetadata &meta,
-                            kernels::signal::resampling::ResamplerCPU &resampler,
+                            kernels::signal::resampling::Resampler &resampler,
                             span<float> decode_scratch_mem, span<float> resample_scratch_mem,
                             float target_sample_rate, bool downmix, const char *audio_filepath);
